+++ scm/domain-base/main/src/api/base/dto/mod.rs
@@ -1,5 +1,9 @@
+pub mod empty_request;
+pub mod empty_response;
 pub mod validation_request;
 pub mod validation_response;
 
+pub use empty_request::EmptyRequest;
+pub use empty_response::EmptyResponse;
 pub use validation_request::ValidationRequest;
 pub use validation_response::ValidationResponse;