+++ scm/domain-base/main/src/api/base/mod.rs
@@ -2,7 +2,7 @@ pub mod dto;
 pub mod errors;
 pub mod traits;
 
-pub use dto::{ValidationRequest, ValidationResponse};
+pub use dto::{EmptyRequest, EmptyResponse, ValidationRequest, ValidationResponse};
 pub use errors::{RequestError, ResponseError};
 pub use traits::Request;
 pub use traits::Response;