+++ scm/domain-base/main/src/api/mod.rs
@@ -1,5 +1,7 @@
 mod base;
 
+pub use base::EmptyRequest;
+pub use base::EmptyResponse;
 pub use base::Request;
 pub use base::RequestError;
 pub use base::Response;